Trait terminfo::capability::Capability

pub trait Capability<'a>: Sized {
    // Required methods
    fn name() -> &'static str;
    fn from(value: Option<&'a Value>) -> Option<Self>;
    fn into(self) -> Option<Value>;
}

A trait for any object that will represent a terminal capability.

Required Methods

fn name() -> &'static str

Returns the name of the capability in its long form.

fn from(value: Option<&'a Value>) -> Option<Self>

Parse the capability from its raw value.

fn into(self) -> Option<Value>

Convert the capability into its raw value.

Object Safety

This trait is not object safe.